dailybuild: suppress warnings
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7983 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWSkeletonSensor.h
@@ -80,12 +80,12 @@ public:
 	}
 	virtual ObjectIf*   GetChildObject(size_t pos)  {
 		if (sensor==NULL) {
-			return GetSkeleton(pos);
+			return GetSkeleton((int)pos);
 		} else {
 			if (pos==0) {
 				return sensor;
 			} else {
-				return GetSkeleton(pos-1);
+				return GetSkeleton((int)pos-1);
 			}
 		}
 	}
